+++ src/py42/clients/authority.py
@@ -2,13 +2,5 @@ from collections import namedtuple
 
 AuthorityClient = namedtuple(
     "AuthorityClient",
-    [
-        "administration",
-        "archive",
-        "devices",
-        "legalhold",
-        "orgs",
-        "securitydata",
-        "users",
-    ],
+    ["administration", "archive", "devices", "legalhold", "orgs", "users"],
 )