+++ src/py42/sdk/__init__.py
@@ -34,7 +34,6 @@ from py42.services.legalhold import LegalHoldService
 from py42.services.orgs import OrgService
 from py42.services.preservationdata import PreservationDataService
 from py42.services.savedsearch import SavedSearchService
-from py42.services.securitydata import SecurityDataService
 from py42.services.storage._service_factory import ConnectionManager
 from py42.services.storage._service_factory import StorageServiceFactory
 from py42.services.users import UserService
@@ -170,7 +169,7 @@ class SDKClient:
         web-restores or finding a file on an archive.
 
         Returns:
-            :class:`py42.services.archive.ArchiveClient`
+            :class:`py42.clients.archive.ArchiveClient`
         """
         return self._clients.archive
 
@@ -231,7 +230,7 @@ class SDKClient:
             * Security plan information
 
         Returns:
-            :class:`py42.services.securitydata.SecurityDataClient`
+            :class:`py42.clients.securitydata.SecurityDataClient`
         """
         return self._clients.securitydata
 
@@ -241,7 +240,7 @@ class SDKClient:
         lists, such as departing employees.
 
         Returns:
-            :class:`py42.services.detectionlists.DetectionListsClient`
+            :class:`py42.clients.detectionlists.DetectionListsClient`
         """
         return self._clients.detectionlists
 
@@ -250,7 +249,7 @@ class SDKClient:
         """A collection of methods related to retrieving and updating alerts rules.
 
         Returns:
-            :class:`py42.services.alertrules.AlertRulesClient`
+            :class:`py42.clients.alertrules.AlertRulesClient`
         """
         return self._clients.alerts
 
@@ -259,7 +258,7 @@ class SDKClient:
         """A collection of methods for retrieving audit logs.
 
         Returns:
-            :class:`py42.services.auditlogs.AuditLogsService`
+            :class:`py42.clients.auditlogs.AuditLogsService`
         """
         return self._clients.auditlogs
 
@@ -318,7 +317,6 @@ def _init_services(main_connection, main_auth):
         devices=DeviceService(main_connection),
         legalhold=LegalHoldService(main_connection),
         orgs=OrgService(main_connection),
-        securitydata=SecurityDataService(main_connection),
         users=UserService(main_connection),
         alertrules=AlertRulesService(alert_rules_conn, user_ctx, user_profile_svc),
         alerts=AlertService(alerts_conn, user_ctx),
@@ -345,7 +343,6 @@ def _init_clients(services, connection):
         devices=services.devices,
         legalhold=services.legalhold,
         orgs=services.orgs,
-        securitydata=services.securitydata,
         users=services.users,
     )
     detectionlists = DetectionListsClient(
@@ -356,7 +353,6 @@ def _init_clients(services, connection):
     )
     alertrules = AlertRulesClient(services.alerts, services.alertrules)
     securitydata = SecurityDataClient(
-        services.securitydata,
         services.fileevents,
         services.preservationdata,
         services.savedsearch,