+++ tests/clients/test_securitydata.py
@@ -1,7 +1,6 @@
 import pytest
 from tests.conftest import create_mock_response
 
-from py42.clients.securitydata import PlanStorageInfo
 from py42.clients.securitydata import SecurityDataClient
 from py42.exceptions import Py42ChecksumNotFoundError
 from py42.exceptions import Py42Error
@@ -10,161 +9,14 @@ from py42.services._connection import Connection
 from py42.services.fileevent import FileEventService
 from py42.services.preservationdata import PreservationDataService
 from py42.services.savedsearch import SavedSearchService
-from py42.services.securitydata import SecurityDataService
 from py42.services.storage._service_factory import StorageServiceFactory
 from py42.services.storage.exfiltrateddata import ExfiltratedDataService
 from py42.services.storage.preservationdata import StoragePreservationDataService
-from py42.services.storage.securitydata import StorageSecurityDataService
 
 FILE_EVENT_URI = "/forensic-search/queryservice/api/v1/fileevent"
 RAW_QUERY = "RAW JSON QUERY"
 USER_UID = "user-uid"
 PDS_EXCEPTION_MESSAGE = "No file with hash {0} available for download."
-GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_ONE_LOCATION = """{
-        "securityPlanLocationsByDestination": [
-            {
-                "destinationGuid": "4",
-                "securityPlanLocationsByNode": [
-                    {
-                        "nodeGuid": "41",
-                        "securityPlanUids": [
-                            "111111111111111111"
-                        ]
-                    }
-                ]
-            }
-        ],
-        "userUid": "917354657784339860"
-    }"""
-GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_TWO_PLANS_ONE_NODE = """{
-        "securityPlanLocationsByDestination": [
-            {
-                "destinationGuid": "4",
-                "securityPlanLocationsByNode": [
-                    {
-                        "nodeGuid": "41",
-                        "securityPlanUids": [
-                            "111111111111111111",
-                            "222222222222222222"
-                        ]
-                    }
-                ]
-            }
-        ],
-        "userUid": "917354657784339860"
-}"""
-GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_TWO_PLANS_TWO_NODES = """{
-        "securityPlanLocationsByDestination": [
-            {
-                "destinationGuid": "4",
-                "securityPlanLocationsByNode": [
-                    {
-                        "nodeGuid": "41",
-                        "securityPlanUids": [
-                            "111111111111111111"
-                        ]
-                    },
-                    {
-                        "nodeGuid": "42",
-                        "securityPlanUids": [
-                            "222222222222222222"
-                        ]
-                    }
-                ]
-            }
-        ],
-        "userUid": "917354657784339860"
-}"""
-GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_ONE_PLAN_TWO_DESTINATIONS = """{
-        "securityPlanLocationsByDestination": [
-            {
-                "destinationGuid": "4",
-                "securityPlanLocationsByNode": [
-                    {
-                        "nodeGuid": "41",
-                        "securityPlanUids": [
-                            "111111111111111111"
-                        ]
-                    }
-                ]
-            },
-            {
-                "destinationGuid": "5",
-                "securityPlanLocationsByNode": [
-                    {
-                        "nodeGuid": "51",
-                        "securityPlanUids": [
-                            "111111111111111111"
-                        ]
-                    }
-                ]
-            }
-        ],
-        "userUid": "917354657784339860"
-}"""
-GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_TWO_PLANS_TWO_DESTINATIONS = """{
-        "securityPlanLocationsByDestination": [
-            {
-                "destinationGuid": "4",
-                "securityPlanLocationsByNode": [
-                    {
-                        "nodeGuid": "41",
-                        "securityPlanUids": [
-                            "111111111111111111",
-                            "222222222222222222"
-                        ]
-                    }
-                ]
-            },
-            {
-                "destinationGuid": "5",
-                "securityPlanLocationsByNode": [
-                    {
-                        "nodeGuid": "51",
-                        "securityPlanUids": [
-                            "111111111111111111",
-                            "222222222222222222"
-                        ]
-                    }
-                ]
-            }
-        ],
-        "userUid": "917354657784339860"
-}"""
-GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_TWO_PLANS_TWO_DESTINATIONS_THREE_NODES = """{
-        "securityPlanLocationsByDestination": [
-            {
-                "destinationGuid": "4",
-                "securityPlanLocationsByNode": [
-                    {
-                        "nodeGuid": "41",
-                        "securityPlanUids": [
-                            "111111111111111111",
-                            "222222222222222222"
-                        ]
-                    }
-                ]
-            },
-            {
-                "destinationGuid": "5",
-                "securityPlanLocationsByNode": [
-                    {
-                        "nodeGuid": "51",
-                        "securityPlanUids": [
-                            "111111111111111111"
-                        ]
-                    },
-                    {
-                        "nodeGuid": "52",
-                        "securityPlanUids": [
-                            "222222222222222222"
-                        ]
-                    }
-                ]
-            }
-        ],
-        "userUid": "917354657784339860"
-}"""
 FILE_EVENTS_RESPONSE = """{
     "fileEvents":[
         {
@@ -306,62 +158,6 @@ class TestSecurityClient:
         return mocker.MagicMock(spec=Connection)
 
     @pytest.fixture
-    def security_service(self, mocker):
-        return mocker.MagicMock(spec=SecurityDataService)
-
-    @pytest.fixture
-    def security_service_one_location(self, mocker, security_service):
-        response = create_mock_response(
-            mocker, GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_ONE_LOCATION
-        )
-        security_service.get_security_event_locations.return_value = response
-        return security_service
-
-    @pytest.fixture
-    def security_service_two_plans_one_node(self, security_service, mocker):
-        response = create_mock_response(
-            mocker, GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_TWO_PLANS_ONE_NODE
-        )
-        security_service.get_security_event_locations.return_value = response
-        return security_service
-
-    @pytest.fixture
-    def security_service_two_plans_two_nodes(self, security_service, mocker):
-        response = create_mock_response(
-            mocker, GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_TWO_PLANS_TWO_NODES
-        )
-        security_service.get_security_event_locations.return_value = response
-        return security_service
-
-    @pytest.fixture
-    def security_service_one_plan_two_destinations(self, security_service, mocker):
-        response = create_mock_response(
-            mocker, GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_ONE_PLAN_TWO_DESTINATIONS
-        )
-        security_service.get_security_event_locations.return_value = response
-        return security_service
-
-    @pytest.fixture
-    def security_service_two_plans_two_destinations(self, security_service, mocker):
-        response = create_mock_response(
-            mocker,
-            GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_TWO_PLANS_TWO_DESTINATIONS,
-        )
-        security_service.get_security_event_locations.return_value = response
-        return security_service
-
-    @pytest.fixture
-    def security_service_two_plans_two_destinations_three_nodes(
-        self, security_service, mocker
-    ):
-        response = create_mock_response(
-            mocker,
-            GET_SECURITY_EVENT_LOCATIONS_RESPONSE_BODY_TWO_PLANS_TWO_DESTINATIONS_THREE_NODES,
-        )
-        security_service.get_security_event_locations.return_value = response
-        return security_service
-
-    @pytest.fixture
     def storage_service_factory(self, mocker):
         return mocker.MagicMock(spec=StorageServiceFactory)
 
@@ -379,14 +175,12 @@ class TestSecurityClient:
 
     def test_search_with_only_query_calls_through_to_client(
         self,
-        security_service,
         file_event_service,
         preservation_data_service,
         saved_search_service,
         storage_service_factory,
     ):
         security_client = SecurityDataClient(
-            security_service,
             file_event_service,
             preservation_data_service,
             saved_search_service,
@@ -395,441 +189,14 @@ class TestSecurityClient:
         security_client.search_file_events(RAW_QUERY)
         file_event_service.search.assert_called_once_with(RAW_QUERY)
 
-    def test_get_security_plan_storage_info_one_location_returns_location_info(
-        self,
-        security_service_one_location,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        security_client = SecurityDataClient(
-            security_service_one_location,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        storage_infos = security_client.get_security_plan_storage_info_list("foo")
-        assert len(storage_infos) == 1
-        assert self._storage_info_contains(
-            storage_infos, "111111111111111111", "4", "41"
-        )
-
-    def test_get_security_plan_storage_info_two_plans_one_node_returns_both_location_info(
-        self,
-        security_service_two_plans_one_node,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        security_client = SecurityDataClient(
-            security_service_two_plans_one_node,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        storage_infos = security_client.get_security_plan_storage_info_list("foo")
-        assert len(storage_infos) == 2
-        assert self._storage_info_contains(
-            storage_infos, "111111111111111111", "4", "41"
-        )
-        assert self._storage_info_contains(
-            storage_infos, "222222222222222222", "4", "41"
-        )
-
-    def test_get_security_plan_storage_info_two_plans_two_nodes_returns_both_location_info(
-        self,
-        security_service_two_plans_two_nodes,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        security_client = SecurityDataClient(
-            security_service_two_plans_two_nodes,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        storage_infos = security_client.get_security_plan_storage_info_list("foo")
-        assert self._storage_info_contains(
-            storage_infos, "111111111111111111", "4", "41"
-        )
-        assert self._storage_info_contains(
-            storage_infos, "222222222222222222", "4", "42"
-        )
-
-    def test_get_security_plan_storage_info_one_plan_two_destinations_returns_one_location(
-        self,
-        security_service_one_plan_two_destinations,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        security_client = SecurityDataClient(
-            security_service_one_plan_two_destinations,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        storage_infos = security_client.get_security_plan_storage_info_list("foo")
-        assert len(storage_infos) == 1
-        assert self._storage_info_contains(
-            storage_infos, "111111111111111111", "4", "41"
-        ) or self._storage_info_contains(storage_infos, "111111111111111111", "5", "51")
-
-    def test_get_security_plan_storage_info_two_plans_two_destinations_returns_one_location_per_plan(
-        self,
-        security_service_two_plans_two_destinations,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        security_client = SecurityDataClient(
-            security_service_two_plans_two_destinations,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        storage_infos = security_client.get_security_plan_storage_info_list("foo")
-        assert len(storage_infos) == 2
-        assert self._storage_info_contains(
-            storage_infos, "111111111111111111", "4", "41"
-        ) or self._storage_info_contains(storage_infos, "111111111111111111", "5", "51")
-        assert self._storage_info_contains(
-            storage_infos, "222222222222222222", "4", "41"
-        ) or self._storage_info_contains(storage_infos, "222222222222222222", "5", "51")
-
-    def test_get_security_plan_storage_info_two_plans_two_destinations_three_nodes_returns_one_location_per_plan(
-        self,
-        security_service_two_plans_two_destinations_three_nodes,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        security_client = SecurityDataClient(
-            security_service_two_plans_two_destinations_three_nodes,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        storage_infos = security_client.get_security_plan_storage_info_list("foo")
-        assert self._storage_info_contains(
-            storage_infos, "111111111111111111", "4", "41"
-        ) or self._storage_info_contains(storage_infos, "111111111111111111", "5", "51")
-        assert self._storage_info_contains(
-            storage_infos, "222222222222222222", "4", "41"
-        ) or self._storage_info_contains(storage_infos, "222222222222222222", "5", "52")
-
-    def test_get_all_user_security_events_calls_security_service_with_expected_params(
-        self,
-        mocker,
-        security_service_one_location,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        mock_storage_security_service = mocker.MagicMock(
-            spec=StorageSecurityDataService
-        )
-        response = create_mock_response(mocker, "{}")
-        mock_storage_security_service.get_plan_security_events.return_value = response
-        storage_service_factory.create_security_data_service.return_value = (
-            mock_storage_security_service
-        )
-        security_client = SecurityDataClient(
-            security_service_one_location,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        for _, _ in security_client.get_all_user_security_events("foo"):
-            pass
-        mock_storage_security_service.get_plan_security_events.assert_called_once_with(
-            "111111111111111111",
-            cursor=None,
-            event_types=None,
-            include_files=True,
-            max_timestamp=None,
-            min_timestamp=None,
-        )
-
-    def test_get_all_user_security_events_when_cursors_returned_calls_security_service_expected_number_of_times(
-        self,
-        mocker,
-        security_service_one_location,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        mock_storage_security_service = mocker.MagicMock(
-            spec=StorageSecurityDataService
-        )
-        response1 = create_mock_response(mocker, '{"cursor": "1:1"}')
-        response2 = create_mock_response(mocker, "{}")
-        mock_storage_security_service.get_plan_security_events.side_effect = [
-            response1,
-            response2,
-        ]
-        storage_service_factory.create_security_data_service.return_value = (
-            mock_storage_security_service
-        )
-        security_client = SecurityDataClient(
-            security_service_one_location,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        for _, _ in security_client.get_all_user_security_events("foo"):
-            pass
-        assert mock_storage_security_service.get_plan_security_events.call_count == 2
-
-    def test_get_all_user_security_events_when_multiple_plans_returned_calls_security_service_expected_number_of_times(
-        self,
-        mocker,
-        security_service_two_plans_one_node,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        mock_storage_security_service = mocker.MagicMock(
-            spec=StorageSecurityDataService
-        )
-        response = create_mock_response(mocker, "{}")
-        mock_storage_security_service.get_plan_security_events.return_value = response
-        storage_service_factory.create_security_data_service.return_value = (
-            mock_storage_security_service
-        )
-        security_client = SecurityDataClient(
-            security_service_two_plans_one_node,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        for _, _ in security_client.get_all_user_security_events("foo"):
-            pass
-        assert mock_storage_security_service.get_plan_security_events.call_count == 2
-
-    def test_get_all_user_security_events_when_multiple_plans_with_cursors_returned_calls_security_service_expected_number_of_times(
-        self,
-        mocker,
-        security_service_two_plans_one_node,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        mock_storage_security_service = mocker.MagicMock(
-            spec=StorageSecurityDataService
-        )
-        response1 = create_mock_response(mocker, '{"cursor": "1:1"}')
-        response2 = create_mock_response(mocker, "{}")
-        mock_storage_security_service.get_plan_security_events.side_effect = [
-            response1,
-            response2,
-            response1,
-            response2,
-        ]
-        storage_service_factory.create_security_data_service.return_value = (
-            mock_storage_security_service
-        )
-        security_client = SecurityDataClient(
-            security_service_two_plans_one_node,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        for _, _ in security_client.get_all_user_security_events("foo"):
-            pass
-        assert mock_storage_security_service.get_plan_security_events.call_count == 4
-
-    @pytest.mark.parametrize(
-        "plan_storage_info",
-        [
-            PlanStorageInfo("111111111111111111", "41", "4"),
-            (PlanStorageInfo("111111111111111111", "41", "4"),),
-            [PlanStorageInfo("111111111111111111", "41", "4")],
-        ],
-    )
-    def test_get_all_plan_security_events_calls_security_service_with_expected_params(
-        self,
-        mocker,
-        security_service,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-        plan_storage_info,
-    ):
-        mock_storage_security_service = mocker.MagicMock(
-            spec=StorageSecurityDataService
-        )
-        response = create_mock_response(mocker, "{}")
-        mock_storage_security_service.get_plan_security_events.return_value = response
-        storage_service_factory.create_security_data_service.return_value = (
-            mock_storage_security_service
-        )
-        security_client = SecurityDataClient(
-            security_service,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        for _, _ in security_client.get_all_plan_security_events(plan_storage_info):
-            pass
-        mock_storage_security_service.get_plan_security_events.assert_called_once_with(
-            "111111111111111111",
-            cursor=None,
-            event_types=None,
-            include_files=True,
-            max_timestamp=None,
-            min_timestamp=None,
-        )
-
-    def test_get_all_plan_security_events_when_cursors_returned_calls_security_service_expected_number_of_times(
-        self,
-        mocker,
-        security_service,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        mock_storage_security_service = mocker.MagicMock(
-            spec=StorageSecurityDataService
-        )
-        response1 = create_mock_response(mocker, '{"cursor": "1:1"}')
-        response2 = create_mock_response(mocker, "{}")
-        mock_storage_security_service.get_plan_security_events.side_effect = [
-            response1,
-            response2,
-        ]
-        storage_service_factory.create_security_data_service.return_value = (
-            mock_storage_security_service
-        )
-        security_client = SecurityDataClient(
-            security_service,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        for _, _ in security_client.get_all_plan_security_events(
-            PlanStorageInfo("111111111111111111", "41", "4")
-        ):
-            pass
-        assert mock_storage_security_service.get_plan_security_events.call_count == 2
-
-    def test_get_all_plan_security_events_when_multiple_plans_returned_calls_security_service_expected_number_of_times(
-        self,
-        mocker,
-        security_service,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        mock_storage_security_service = mocker.MagicMock(
-            spec=StorageSecurityDataService
-        )
-        response = create_mock_response(mocker, "{}")
-        mock_storage_security_service.get_plan_security_events.return_value = response
-        storage_service_factory.create_security_data_service.return_value = (
-            mock_storage_security_service
-        )
-        security_client = SecurityDataClient(
-            security_service,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        plans = [
-            PlanStorageInfo("111111111111111111", "41", "4"),
-            PlanStorageInfo("222222222222222222", "41", "4"),
-        ]
-        for _, _ in security_client.get_all_plan_security_events(plans):
-            pass
-        assert mock_storage_security_service.get_plan_security_events.call_count == 2
-
-    def test_get_all_plan_security_events_when_multiple_plans_with_cursors_returned_calls_security_service_expected_number_of_times(
-        self,
-        mocker,
-        security_service,
-        file_event_service,
-        preservation_data_service,
-        saved_search_service,
-        storage_service_factory,
-    ):
-        mock_storage_security_service = mocker.MagicMock(
-            spec=StorageSecurityDataService
-        )
-        response1 = create_mock_response(mocker, '{"cursor": "1:1"}')
-        response2 = create_mock_response(mocker, "{}")
-        mock_storage_security_service.get_plan_security_events.side_effect = [
-            response1,
-            response2,
-            response1,
-            response2,
-        ]
-        storage_service_factory.create_security_data_service.return_value = (
-            mock_storage_security_service
-        )
-        security_client = SecurityDataClient(
-            security_service,
-            file_event_service,
-            preservation_data_service,
-            saved_search_service,
-            storage_service_factory,
-        )
-        plans = [
-            PlanStorageInfo("111111111111111111", "41", "4"),
-            PlanStorageInfo("222222222222222222", "41", "4"),
-        ]
-        for _, _ in security_client.get_all_plan_security_events(plans):
-            pass
-        assert mock_storage_security_service.get_plan_security_events.call_count == 4
-
-    # the order the items are iterated through is not deterministic in some versions of python,
-    # so we simply test that the value returned is one of the _possible_ values.
-    def _storage_info_contains(
-        self, storage_info_list, plan_uid, destination_guid, node_guid
-    ):
-        return any(
-            item.plan_uid == plan_uid
-            and item.destination_guid == destination_guid
-            and item.node_guid == node_guid
-            for item in storage_info_list
-        )
-
     def test_saved_searches_returns_saved_search_client(
         self,
-        security_service,
         file_event_service,
         preservation_data_service,
         saved_search_service,
         storage_service_factory,
     ):
         security_client = SecurityDataClient(
-            security_service,
             file_event_service,
             preservation_data_service,
             saved_search_service,
@@ -853,7 +220,6 @@ class TestSecurityClient:
     def pds_config(
         self,
         mocker,
-        security_service,
         storage_service_factory,
         file_event_service,
         preservation_data_service,
@@ -883,7 +249,6 @@ class TestSecurityClient:
             exfiltration_client
         )
 
-        mock.security_service = security_service
         mock.storage_service_factory = storage_service_factory
         mock.file_event_service = file_event_service
         mock.preservation_data_service = preservation_data_service
@@ -897,7 +262,6 @@ class TestSecurityClient:
     ):
 
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -935,7 +299,6 @@ class TestSecurityClient:
             mocker, FILE_EVENTS_RESPONSE.replace("-2", "-6")
         )
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -974,7 +337,6 @@ class TestSecurityClient:
             mocker, '{"fileEvents": []}'
         )
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -999,7 +361,6 @@ class TestSecurityClient:
         )
 
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1032,7 +393,6 @@ class TestSecurityClient:
             file_location
         )
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1059,7 +419,6 @@ class TestSecurityClient:
         )
 
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1075,7 +434,6 @@ class TestSecurityClient:
         self, pds_config,
     ):
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1114,7 +472,6 @@ class TestSecurityClient:
         )
 
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1153,7 +510,6 @@ class TestSecurityClient:
             mocker, '{"fileEvents": []}'
         )
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1178,7 +534,6 @@ class TestSecurityClient:
         )
 
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1211,7 +566,6 @@ class TestSecurityClient:
             file_location
         )
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1238,7 +592,6 @@ class TestSecurityClient:
         )
 
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1259,7 +612,6 @@ class TestSecurityClient:
         )
 
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1286,7 +638,6 @@ class TestSecurityClient:
         )
 
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1313,7 +664,6 @@ class TestSecurityClient:
         )
 
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1340,7 +690,6 @@ class TestSecurityClient:
         )
 
         security_client = SecurityDataClient(
-            pds_config.security_service,
             pds_config.file_event_service,
             pds_config.preservation_data_service,
             pds_config.saved_search_service,
@@ -1360,7 +709,6 @@ class TestSecurityClient:
 
     def test_search_all_file_events_calls_search_with_expected_params_when_pg_token_is_not_passed(
         self,
-        security_service,
         connection,
         preservation_data_service,
         saved_search_service,
@@ -1376,7 +724,6 @@ class TestSecurityClient:
         connection.post.return_value = successful_response
 
         security_client = SecurityDataClient(
-            security_service,
             file_event_service,
             preservation_data_service,
             saved_search_service,
@@ -1397,7 +744,6 @@ class TestSecurityClient:
 
     def test_search_all_file_events_calls_search_with_expected_params_when_pg_token_is_passed(
         self,
-        security_service,
         connection,
         preservation_data_service,
         saved_search_service,
@@ -1412,7 +758,6 @@ class TestSecurityClient:
         }
         connection.post.return_value = successful_response
         security_client = SecurityDataClient(
-            security_service,
             file_event_service,
             preservation_data_service,
             saved_search_service,
@@ -1434,14 +779,12 @@ class TestSecurityClient:
     def test_search_all_file_events_handles_unescaped_quote_chars_in_token(
         self,
         connection,
-        security_service,
         preservation_data_service,
         saved_search_service,
         storage_service_factory,
     ):
         file_event_service = FileEventService(connection)
         security_client = SecurityDataClient(
-            security_service,
             file_event_service,
             preservation_data_service,
             saved_search_service,
@@ -1463,14 +806,12 @@ class TestSecurityClient:
     def test_search_all_file_events_handles_escaped_quote_chars_in_token(
         self,
         connection,
-        security_service,
         preservation_data_service,
         saved_search_service,
         storage_service_factory,
     ):
         file_event_service = FileEventService(connection)
         security_client = SecurityDataClient(
-            security_service,
             file_event_service,
             preservation_data_service,
             saved_search_service,