+++ tests/sdk/test_sdk.py
@@ -7,7 +7,6 @@ from py42.clients.archive import ArchiveClient
 from py42.clients.auditlogs import AuditLogsClient
 from py42.clients.cases import CasesClient
 from py42.clients.detectionlists import DetectionListsClient
-from py42.clients.securitydata import SecurityDataClient
 from py42.exceptions import Py42UnauthorizedError
 from py42.sdk import from_local_account
 from py42.sdk import SDKClient
@@ -70,10 +69,6 @@ class TestSDK:
     def test_has_org_service_set(self, py42_connection, mock_auth):
         client = SDKClient(py42_connection, mock_auth)
         assert type(client.orgs) == orgs.OrgService
-
-    def test_has_security_service_set(self, py42_connection, mock_auth):
-        client = SDKClient(py42_connection, mock_auth)
-        assert type(client.securitydata) == SecurityDataClient
 
     def test_has_user_service_set(self, py42_connection, mock_auth):
         client = SDKClient(py42_connection, mock_auth)